export NewZoneContentForm in own component in forms folder
