+++ src/pages/main/ZonesContent.tsx
@@ -2,22 +2,10 @@ import SkeletonPage from '@/components/features/SkeletonPage'
 import FormDialog from '@/components/shared/FormDialog'
 import { useState } from 'react'
 import { useTranslation } from 'react-i18next'
-import {
-  TextField,
-  Checkbox,
-  FormControlLabel,
-  FormControl,
-  InputLabel,
-  Select,
-  MenuItem,
-  ListItemText,
-  OutlinedInput
-} from '@mui/material'
-import { Controller, UseFormReturn, SubmitHandler } from 'react-hook-form'
+import { SubmitHandler } from 'react-hook-form'
 import { NewZoneFormData, newZoneSchema } from '@/schemas/newZoneSchema'
 import ZonesContentsTable from '@/components/features/main/ZonesContentsTable'
-
-const markers = ['Масло', 'Гуми', 'Чистачки']
+import NewZoneContentForm from '@/components/features/forms/NewZoneContentForm'
 
 export default function ZonesContent() {
   const { t: translate } = useTranslation()
@@ -34,69 +22,6 @@ export default function ZonesContent() {
   const handleSubmit: SubmitHandler<NewZoneFormData> = (data) => {
     console.log(data)
     onCloseDialog()
-  }
-
-  function CreateZoneForm({ control, formState: { errors } }: UseFormReturn<NewZoneFormData>) {
-    return (
-      <>
-        <Controller
-          name="zoneName"
-          control={control}
-          render={({ field }) => (
-            <TextField
-              {...field}
-              label={translate('newZone.labels.name')}
-              id="zoneName"
-              name="zoneName"
-              required
-              fullWidth
-              autoFocus
-              error={!!errors.zoneName}
-              helperText={errors.zoneName?.message ? translate(errors.zoneName.message) : ''}
-            />
-          )}
-        />
-
-        <Controller
-          name="markers"
-          control={control}
-          render={({ field }) => (
-            <FormControl fullWidth>
-              <InputLabel id="demo-multiple-checkbox-label">
-                {translate('newZone.labels.markers')}
-              </InputLabel>
-              <Select
-                {...field}
-                labelId="demo-multiple-checkbox-label"
-                id="demo-multiple-checkbox"
-                multiple
-                value={field.value || []}
-                onChange={(e) => field.onChange(e.target.value)}
-                input={<OutlinedInput />}
-                renderValue={(selected) => (selected as string[]).join(', ')}>
-                {markers.map((marker) => (
-                  <MenuItem key={marker} value={marker}>
-                    <Checkbox checked={field.value?.includes(marker)} />{' '}
-                    <ListItemText primary={marker} />
-                  </MenuItem>
-                ))}
-              </Select>
-            </FormControl>
-          )}
-        />
-
-        <Controller
-          name="isFinal"
-          control={control}
-          render={({ field }) => (
-            <FormControlLabel
-              control={<Checkbox {...field} checked={field.value} />}
-              label={translate('newZone.labels.isFinal')}
-            />
-          )}
-        />
-      </>
-    )
   }
 
   return (
@@ -117,7 +42,7 @@ export default function ZonesContent() {
         onCloseDialog={onCloseDialog}
         schema={newZoneSchema}
         onSubmit={handleSubmit}
-        renderForm={CreateZoneForm}
+        renderForm={(methods) => <NewZoneContentForm {...methods} />}
       />
     </>
   )
